test: enable creation of in-memory sqlite3 databases for testing
I have chosen this approach to improve the portability of the
tests and to allow for parallelization. By using an in-memory database,
I can create multiple SQLite3 databases at the same time without the
need for a real Postgres database. Additionally, I have chosen not to
use the package sequelize-mock as it is outdated.